pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "almagest"
version = "0.1.0"
description = "AI research agent for NASA ADS — multi-workflow CLI powered by Claude"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "schedule>=1.2.0",
    "openai>=1.0.0",       # for local LLM via OpenAI-compatible API
    "pypdf>=4.0.0",        # for local PDF text extraction
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
almagest = "almagest.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/almagest"]

[tool.ruff]
line-length = 100
target-version = "py311"
